+++ link-grammar/build-disjuncts.h
@@ -14,11 +14,11 @@
 #include "api-types.h"
 #include "structures.h"
 
-void build_sentence_disjuncts(Sentence sent, float cost_cutoff);
+void build_sentence_disjuncts(Sentence sent, double cost_cutoff);
 X_node *   build_word_expressions(Dictionary dict, const char *);
 #ifdef OBSOLETE_MEMORY_PIGGY
 Disjunct * build_disjuncts_for_dict_node(Dict_node *);
 #endif
-Disjunct * build_disjuncts_for_X_node(X_node * x, float cost_cutoff);
+Disjunct * build_disjuncts_for_X_node(X_node * x, double cost_cutoff);
 
 unsigned int count_disjunct_for_dict_node(Dict_node *dn);
